use user_id instead of team_id, add new table 'minuteur_taches'

+++ src/API/minuteurs/delete-many.php
@@ -4,8 +4,8 @@ require_once __DIR__ . '/../database.php';
 $data = getPostData();
 $db = openDatabase();
 $placeholders = implode(',', array_fill(0, count($data['ids']), '?'));
-$sql = "DELETE FROM `minuteurs` WHERE `id` IN ($placeholders)";
-exequteSQL($db, $sql, $data['ids']);
+$sql = "DELETE FROM `minuteurs` WHERE `user_id` = ? AND `id` IN ($placeholders)";
+exequteSQL($db, $sql, $data['user'], $data['ids']);
 $db->close();
 echo 'done';
 ?>
